backend/requirements.txt
fastapi
uvicorn[standard]
boto3
httpx
python-multipart
python-dotenv
pydantic
